+++ src/androidMain/kotlin/KronSchedulerWork.kt
@@ -139,7 +139,6 @@ abstract class KronSchedulerWork(
             }
             throw e
         }
-        result.outputData
         when (result) {
             is Result.Failure -> if (prolongOnFailure) prolong()
             is Result.Success -> if (prolongOnSuccess) prolong()